tests: support testing remote taish-server

+++ tests/test_base.py
@@ -5,9 +5,19 @@ import os
 import time
 import taish
 
-TAI_TEST_MODULE_LOCATION = os.environ.get('TAI_TEST_MODULE_LOCATION', '0')
-if TAI_TEST_MODULE_LOCATION == '':
+TAI_TEST_MODULE_LOCATION = os.environ.get('TAI_TEST_MODULE_LOCATION', '')
+if not TAI_TEST_MODULE_LOCATION:
     TAI_TEST_MODULE_LOCATION = '0'
+
+TAI_TEST_TAISH_SERVER_ADDRESS = os.environ.get('TAI_TEST_TAISH_SERVER_ADDRESS', '')
+if not TAI_TEST_TAISH_SERVER_ADDRESS:
+    TAI_TEST_TAISH_SERVER_ADDRESS = taish.DEFAULT_SERVER_ADDRESS
+
+TAI_TEST_TAISH_SERVER_PORT = os.environ.get('TAI_TEST_TAISH_SERVER_PORT', '')
+if not TAI_TEST_TAISH_SERVER_PORT:
+    TAI_TEST_TAISH_SERVER_PORT = taish.DEFAULT_SERVER_PORT
+
+TAI_TEST_NO_LOCAL_TAISH_SERVER = True if os.environ.get('TAI_TEST_NO_LOCAL_TAISH_SERVER', '') else False
 
 def output_reader(proc):
     for line in iter(proc.stdout.readline, b''):
@@ -16,6 +26,8 @@ def output_reader(proc):
 class TestTAI(unittest.TestCase):
 
     def setUp(self):
+        if TAI_TEST_NO_LOCAL_TAISH_SERVER:
+            return
         proc = sp.Popen('taish_server', stderr=sp.STDOUT, stdout=sp.PIPE)
         self.d = threading.Thread(target=output_reader, args=(proc,))
         self.d.start()
@@ -23,14 +35,15 @@ class TestTAI(unittest.TestCase):
         time.sleep(5) # wait for the server to be ready
 
     def tearDown(self):
-        print('tear down')
+        if TAI_TEST_NO_LOCAL_TAISH_SERVER:
+            return
         self.proc.terminate()
         self.proc.wait(timeout=0.2)
         self.d.join()
         self.proc.stdout.close()
 
     def test_list(self):
-        cli = taish.Client()
+        cli = taish.Client(TAI_TEST_TAISH_SERVER_ADDRESS, TAI_TEST_TAISH_SERVER_PORT)
         m = cli.list()
         self.assertNotEqual(m, None)
         self.assertTrue(TAI_TEST_MODULE_LOCATION in m)
@@ -38,13 +51,13 @@ class TestTAI(unittest.TestCase):
         print('module oid: 0x{:x}'.format(module.oid))
 
     def test_get_module(self):
-        cli = taish.Client()
+        cli = taish.Client(TAI_TEST_TAISH_SERVER_ADDRESS, TAI_TEST_TAISH_SERVER_PORT)
         m = cli.get_module(TAI_TEST_MODULE_LOCATION)
         self.assertNotEqual(m, None)
         print('module oid: 0x{:x}'.format(m.oid))
 
     def test_get_netif(self):
-        cli = taish.Client()
+        cli = taish.Client(TAI_TEST_TAISH_SERVER_ADDRESS, TAI_TEST_TAISH_SERVER_PORT)
         m = cli.get_module(TAI_TEST_MODULE_LOCATION)
         self.assertNotEqual(m, None)
         netif = m.get_netif()
@@ -52,7 +65,7 @@ class TestTAI(unittest.TestCase):
         print('netif oid: 0x{:x}'.format(netif.oid))
 
     def test_get_hostif(self):
-        cli = taish.Client()
+        cli = taish.Client(TAI_TEST_TAISH_SERVER_ADDRESS, TAI_TEST_TAISH_SERVER_PORT)
         m = cli.get_module(TAI_TEST_MODULE_LOCATION)
         self.assertNotEqual(m, None)
         hostif = m.get_hostif()
@@ -60,7 +73,7 @@ class TestTAI(unittest.TestCase):
         print('hostif oid: 0x{:x}'.format(hostif.oid))
 
     def test_module(self):
-        cli = taish.Client()
+        cli = taish.Client(TAI_TEST_TAISH_SERVER_ADDRESS, TAI_TEST_TAISH_SERVER_PORT)
         m = cli.get_module(TAI_TEST_MODULE_LOCATION)
         for metadata in m.list_attribute_metadata():
             try:
@@ -70,7 +83,7 @@ class TestTAI(unittest.TestCase):
             print('{}: {}'.format(metadata.short_name, v))
 
     def test_netif(self):
-        cli = taish.Client()
+        cli = taish.Client(TAI_TEST_TAISH_SERVER_ADDRESS, TAI_TEST_TAISH_SERVER_PORT)
         m = cli.get_module(TAI_TEST_MODULE_LOCATION)
         netif = m.get_netif()
         for metadata in netif.list_attribute_metadata():
@@ -81,7 +94,7 @@ class TestTAI(unittest.TestCase):
             print('{}: {}'.format(metadata.short_name, v))
 
     def test_netif_set_output_power(self):
-        cli = taish.Client()
+        cli = taish.Client(TAI_TEST_TAISH_SERVER_ADDRESS, TAI_TEST_TAISH_SERVER_PORT)
         m = cli.get_module(TAI_TEST_MODULE_LOCATION)
         netif = m.get_netif()
         netif.set('output-power', '-4')
@@ -90,7 +103,7 @@ class TestTAI(unittest.TestCase):
         self.assertEqual(round(float(netif.get('output-power'))), -5)
 
     def test_netif_set_modulation_format(self):
-        cli = taish.Client()
+        cli = taish.Client(TAI_TEST_TAISH_SERVER_ADDRESS, TAI_TEST_TAISH_SERVER_PORT)
         m = cli.get_module(TAI_TEST_MODULE_LOCATION)
         netif = m.get_netif()
         netif.set('modulation-format', 'dp-64-qam')
@@ -99,7 +112,7 @@ class TestTAI(unittest.TestCase):
         self.assertEqual(netif.get('modulation-format'), 'dp-qpsk')
 
     def test_hostif(self):
-        cli = taish.Client()
+        cli = taish.Client(TAI_TEST_TAISH_SERVER_ADDRESS, TAI_TEST_TAISH_SERVER_PORT)
         m = cli.get_module(TAI_TEST_MODULE_LOCATION)
         hostif = m.get_hostif()
         for metadata in hostif.list_attribute_metadata():
@@ -110,7 +123,7 @@ class TestTAI(unittest.TestCase):
             print('{}: {}'.format(metadata.short_name, v))
 
     def test_hostif_set_fec(self):
-        cli = taish.Client()
+        cli = taish.Client(TAI_TEST_TAISH_SERVER_ADDRESS, TAI_TEST_TAISH_SERVER_PORT)
         m = cli.get_module(TAI_TEST_MODULE_LOCATION)
         hostif = m.get_hostif()
         hostif.set('fec-type', 'rs')
@@ -121,7 +134,7 @@ class TestTAI(unittest.TestCase):
         self.assertEqual(hostif.get('fec-type'), 'none')
 
     def test_hostif_set_loopback(self):
-        cli = taish.Client()
+        cli = taish.Client(TAI_TEST_TAISH_SERVER_ADDRESS, TAI_TEST_TAISH_SERVER_PORT)
         m = cli.get_module(TAI_TEST_MODULE_LOCATION)
         hostif = m.get_hostif()
         hostif.set('loopback-type', 'shallow')
@@ -132,7 +145,7 @@ class TestTAI(unittest.TestCase):
         self.assertEqual(hostif.get('loopback-type'), 'none')
 
     def test_remove(self):
-        cli = taish.Client()
+        cli = taish.Client(TAI_TEST_TAISH_SERVER_ADDRESS, TAI_TEST_TAISH_SERVER_PORT)
         l = cli.list()
         self.assertNotEqual(l, None)
         self.assertTrue(TAI_TEST_MODULE_LOCATION in l)
@@ -167,7 +180,7 @@ class TestTAI(unittest.TestCase):
 
     def test_create(self):
         self.test_remove()
-        cli = taish.Client()
+        cli = taish.Client(TAI_TEST_TAISH_SERVER_ADDRESS, TAI_TEST_TAISH_SERVER_PORT)
         with self.assertRaises(Exception):
             cli.get_module(TAI_TEST_MODULE_LOCATION)
         cli.create('module', [('location', TAI_TEST_MODULE_LOCATION)])
